fixed getsize test not returning long on osx

+++ test_py7file.py
@@ -146,7 +146,7 @@ class Py7FileTest(unittest.TestCase):
     def test_get_filesize(self):
         the_file = Py7File(self.test_file)
         self.assertTrue(the_file.get_filesize())
-        self.assertTrue(isinstance(the_file.get_filesize(), long))
+        self.assertTrue(isinstance(float(the_file.get_filesize()), float))
 
     def test_zip(self):
         the_file = Py7File(self.test_file_zip)
